prevent integer overflow (though unlikely in large network)
git-svn-id: file:///home/toad/git-migration/temprepository/trunk/freenet@14322 67a373e5-eb02-0410-a15c-ee090a768436

+++ src/freenet/node/OpennetManager.java
@@ -40,7 +40,7 @@ public class OpennetManager {
 	/** Time at which last dropped a peer */
 	private long timeLastDropped;
 	/** Number of successful CHK requests since last added a node */
-	private int successCount;
+	private long successCount;
 	
 	/** Only drop a connection after at least this many successful requests */
 	static final int MIN_SUCCESS_BETWEEN_DROP_CONNS = 10;
